scraper/requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pyyaml==6.0.2
pandas==2.2.3
matplotlib==3.9.3
seaborn==0.13.2
jupyter==1.1.1
pytest==8.3.4
responses==0.25.6
